+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2019-08-17
+- Added `getLowestFitnessIndividual` and `getHighestFitnessIndividual` functions to `Genemo.run` result.
+
 ## 1.1.0 - 2019-08-02
 - Added `Genemo.crossover.edgeRecombination` function.
 - Added type checking to `elitism`, `evaluatePopulation`, `generateInitialPopulation`, `stopCondition` and all selection functions.